Split indexing.py
